[lldb][PlatformDarwin][test] Move Platform test utilities into common header for re-use
In https://github.com/llvm/llvm-project/pull/187031 I'm planning on
re-using the `MockScriptInterpreterPython` and `CreateFile` API from a
different unit-test in `Platform/`.
